# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meditiya Sathi - Digital Society Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, past, all]
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Register for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRegistrationInput"
      responses:
        "201":
          description: Registration confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistration"
...
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List donations
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
    post:
      operationId: createDonation
      tags: [donations]
      summary: Record a donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
  /donations/top-donors:
    get:
      operationId: getTopDonors
      tags: [donations]
      summary: Get top donors leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top donors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DonorSummary"

  # ── FESTIVALS ──────────────────────────────────────────────────────────────
  /festivals:
    get:
      operationId: listFestivals
      tags: [festivals]
      summary: List festivals
      responses:
        "200":
          description: Festivals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Festival"
    post:
      operationId: createFestival
      tags: [festivals]
      summary: Create festival (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FestivalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Festival"
  /festivals/{slug}:
    get:
      operationId: getFestival
      tags: [festivals]
      summary: Get festival details by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Festival
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Festival"

  # ── VOLUNTEERS ─────────────────────────────────────────────────────────────
...
  /competitions:
    get:
      operationId: listCompetitions
      tags: [competitions]
      summary: List competitions
      responses:
        "200":
          description: Competitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competition"
    post:
      operationId: createCompetition
      tags: [competitions]
      summary: Create competition (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
  /competitions/{id}/register:
    post:
      operationId: registerForCompetition
      tags: [competitions]
      summary: Register for a competition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionRegistrationInput"
      responses:
        "201":
          description: Registration confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionRegistration"
  /competitions/{id}/winners:
    get:
      operationId: getCompetitionWinners
      tags: [competitions]
      summary: Get winners for a competition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Winners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompetitionWinner"

  # ── COMMUNITY SERVICES ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, title, description, date, location, category, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        location:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        maxParticipants:
          type: ["integer", "null"]
        registrationCount:
          type: integer
        festivalId:
          type: ["integer", "null"]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, description, date, location, category]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
        location:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        maxParticipants:
          type: ["integer", "null"]
        festivalId:
          type: ["integer", "null"]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
        maxParticipants:
          type: ["integer", "null"]
    EventRegistration:
      type: object
      required: [id, eventId, userId, registeredAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: string
        userName:
          type: string
        registeredAt:
          type: string
    EventRegistrationInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        familyMembers:
          type: ["integer", "null"]

    # ── NOTICES ─────────────────────────────────────────────────────────────
...
    Donation:
      type: object
      required: [id, amount, donorName, festivalId, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        donorName:
          type: string
        donorPhone:
          type: ["string", "null"]
        festivalId:
          type: ["integer", "null"]
        transactionId:
          type: ["string", "null"]
        method:
          type: string
          enum: [upi, cash, online]
        message:
          type: ["string", "null"]
        createdAt:
          type: string
    DonationInput:
      type: object
      required: [amount, donorName, method]
      properties:
        amount:
          type: number
        donorName:
          type: string
        donorPhone:
          type: ["string", "null"]
        festivalId:
          type: ["integer", "null"]
        transactionId:
          type: ["string", "null"]
        method:
          type: string
          enum: [upi, cash, online]
        message:
          type: ["string", "null"]
    DonorSummary:
      type: object
      required: [donorName, totalAmount, donationCount]
      properties:
        donorName:
          type: string
        totalAmount:
          type: number
        donationCount:
          type: integer

    # ── FESTIVALS ───────────────────────────────────────────────────────────
    Festival:
      type: object
      required: [id, name, slug, description, year, startDate, endDate]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        year:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        bannerImageUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        history:
          type: ["string", "null"]
        isActive:
          type: boolean
    FestivalInput:
      type: object
      required: [name, slug, description, year, startDate, endDate]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        year:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        bannerImageUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        history:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── VOLUNTEERS ──────────────────────────────────────────────────────────
...
    Competition:
      type: object
      required: [id, name, category, description, date, status]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [drawing, rangoli, dance, singing, chess, cricket, indoor_games, quiz]
        description:
          type: string
        date:
          type: string
          format: date-time
        ageGroup:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        prizes:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        maxParticipants:
          type: ["integer", "null"]
        registrationCount:
          type: integer
        festivalId:
          type: ["integer", "null"]
    CompetitionInput:
      type: object
      required: [name, category, description, date]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        date:
          type: string
        ageGroup:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        prizes:
          type: ["string", "null"]
        maxParticipants:
          type: ["integer", "null"]
        festivalId:
          type: ["integer", "null"]
    CompetitionRegistration:
      type: object
      required: [id, competitionId, participantName, registeredAt]
      properties:
        id:
          type: integer
        competitionId:
          type: integer
        participantName:
          type: string
        phone:
          type: string
        ageGroup:
          type: ["string", "null"]
        registeredAt:
          type: string
    CompetitionRegistrationInput:
      type: object
      required: [participantName, phone]
      properties:
        participantName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        ageGroup:
          type: ["string", "null"]
    CompetitionWinner:
      type: object
      required: [id, competitionId, participantName, position]
      properties:
        id:
          type: integer
        competitionId:
          type: integer
        participantName:
          type: string
        position:
          type: integer
        prize:
          type: ["string", "null"]

    # ── COMMUNITY SERVICES ──────────────────────────────────────────────────
...